Only record current group

+++ scripts/components/interiorvisitor.lua
@@ -369,44 +369,48 @@ function InteriorVisitor:RevealAlwaysShownMinimapEntities()
     local interior_group = self.center_ent:GetGroupId()
 
     for ent in pairs(TheWorld.components.interiormaprevealer.tracking_entities) do
-        local pos = ent:GetPosition()
-        local center = TheWorld.components.interiorspawner:GetInteriorCenter(pos)
-        local current_data = self.always_shown_minimap_entities[ent]
-        if current_data then
-            if interior_group ~= center:GetGroupId() then
-                table.insert(sync_actions, {
-                    type = "delete",
-                    data = ent.Network:GetNetworkID(),
-                })
-                self.always_shown_minimap_entities[ent] = nil
-            else
-                local offset = pos - center:GetPosition()
-                if current_data.offset_x ~= offset.x or current_data.z ~= offset.z then
-                    current_data.offset_x = offset.x
-                    current_data.offset_z = offset.z
+        local network_id = ent.Network and ent.Network:GetNetworkID()
+        -- Some mods have entities with minimap icon but without network
+        if network_id then
+            local pos = ent:GetPosition()
+            local center = TheWorld.components.interiorspawner:GetInteriorCenter(pos)
+            local current_data = self.always_shown_minimap_entities[ent]
+            if current_data then
+                if interior_group ~= center:GetGroupId() then
                     table.insert(sync_actions, {
-                        type = "replace",
-                        data = current_data,
+                        type = "delete",
+                        data = network_id,
                     })
+                    self.always_shown_minimap_entities[ent] = nil
+                else
+                    local offset = pos - center:GetPosition()
+                    if current_data.offset_x ~= offset.x or current_data.z ~= offset.z then
+                        current_data.offset_x = offset.x
+                        current_data.offset_z = offset.z
+                        table.insert(sync_actions, {
+                            type = "replace",
+                            data = current_data,
+                        })
+                    end
                 end
+            elseif interior_group == center:GetGroupId() then
+                local offset = pos - center:GetPosition()
+                local coord_x, coord_y = center:GetCoordinates()
+                current_data = {
+                    id = network_id,
+                    coord_x = coord_x,
+                    coord_y = coord_y,
+                    offset_x = offset.x,
+                    offset_z = offset.z,
+                    icon = ent.MiniMapEntity:GetIcon(),
+                    priority = ent.MiniMapEntity:GetPriority(),
+                }
+                self.always_shown_minimap_entities[ent] = current_data
+                table.insert(sync_actions, {
+                    type = "replace",
+                    data = current_data,
+                })
             end
-        else
-            local offset = pos - center:GetPosition()
-            local coord_x, coord_y = center:GetCoordinates()
-            current_data = {
-                id = ent.Network:GetNetworkID(),
-                coord_x = coord_x,
-                coord_y = coord_y,
-                offset_x = offset.x,
-                offset_z = offset.z,
-                icon = ent.MiniMapEntity:GetIcon(),
-                priority = ent.MiniMapEntity:GetPriority(),
-            }
-            self.always_shown_minimap_entities[ent] = current_data
-            table.insert(sync_actions, {
-                type = "replace",
-                data = current_data,
-            })
         end
     end
